+++ src/main/java/com/demo/model/ParkingEvent.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 
@@ -16,13 +17,13 @@ public class ParkingEvent {
     @Id
     @GeneratedValue
     private long Id;
-    private double paymentPln;
+    private BigDecimal paymentPln;
     protected Timestamp paymentDate;
 
     public ParkingEvent() {
     }
 
-    public ParkingEvent(double paymentPln, Timestamp paymentDate) {
+    public ParkingEvent(BigDecimal paymentPln, Timestamp paymentDate) {
         this.paymentPln = paymentPln;
         this.paymentDate = paymentDate;
     }
